При обновлении в уже запущенном tmux выключает его, добавил проверку работу под tmux

+++ zsh-update-configs.sh
@@ -59,14 +59,21 @@ if [ ! -d ~/.tmux/plugins/tpm ]; then
 	git pull
 	cd -
 fi
-# start a server but don't attach to it
-tmux start-server
-# create a new session but don't attach to it either
-tmux new-session -d
-# install the plugins
-~/.tmux/plugins/tpm/scripts/install_plugins.sh
-# killing the server is not required, I guess
-tmux kill-server
+
+if [ -n "$TMUX" ]; then
+	# install the plugins
+	~/.tmux/plugins/tpm/scripts/install_plugins.sh
+else
+	# start a server but don't attach to it
+	tmux start-server
+	# create a new session but don't attach to it either
+	tmux new-session -d
+	# install the plugins
+	~/.tmux/plugins/tpm/scripts/install_plugins.sh
+	# killing the server is not required, I guess
+	tmux kill-server
+fi
+
 
 unset SCRIPT
 unset BASEDIR
